refactors environment out
