+++ src/lib.rs
@@ -1,5 +1,6 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
+pub mod environment;
 pub mod font;
 pub mod layout;
 pub mod primitives;
